Refactor Value API and add GenericBaseValue
