+++ src/main/java/mchorse/mclib/config/values/GenericValue.java
@@ -30,11 +30,9 @@ import javax.annotation.Nullable;
  * @param <T> the datatype of the values in this value container
  * @author Christian F (Chryfi)
  */
-public abstract class GenericValue<T> extends Value
+public abstract class GenericValue<T> extends GenericBaseValue<T>
 {
-    protected T value;
     protected T defaultValue;
-    protected T serverValue;
 
     public GenericValue(String id)
     {
@@ -58,157 +56,12 @@ public abstract class GenericValue<T> extends Value
     }
 
     /**
-     * @return reference to {@link #value}, or if the {@link #serverValue} != null return {@link #serverValue}.
-     *         If the generic datatype is instance of {@link ICopy}, the return value will be copied.
-     */
-    public T get()
-    {
-        if (this.serverValue == null)
-        {
-            if (this.value instanceof ICopy)
-            {
-                return ((ICopy<T>) this.value).copy();
-            }
-            else
-            {
-                return this.value;
-            }
-        }
-        else
-        {
-            if (this.serverValue instanceof ICopy)
-            {
-                return ((ICopy<T>) this.serverValue).copy();
-            }
-            else
-            {
-                return this.serverValue;
-            }
-        }
-    }
-
-    /**
-     * Set the value of this instance to the provided value. If the value is instanceOf {@link ICopy}, it will be copied.
-     * If the value is null, the result of {@link #getNullValue()} will be assigned.
-     * This method calls {@link #saveLater()}
-     * @param value
-     */
-    public void set(T value)
-    {
-        if (value == null)
-        {
-            this.value = this.getNullValue();
-        }
-        else
-        {
-            if (value instanceof ICopy)
-            {
-                this.value = ((ICopy<T>) value).copy();
-            }
-            else
-            {
-                this.value = value;
-            }
-        }
-
-        this.saveLater();
-    }
-
-    /**
      * Reset this value to defaultValue. Calls {@link #set(Object)}
      */
     @Override
     public void reset()
     {
         this.set(this.defaultValue);
-    }
-
-    /**
-     * Only used in Aperture undo/redo system
-     * @param value if the specified object is not null and assignable to the generic type T,
-     *             the value be set, using the {@link #set(Object)} method.
-     */
-    @Override
-    public void setValue(Object value)
-    {
-        if (value == null)
-        {
-            return;
-        }
-
-        try
-        {
-            this.set((T) value);
-        }
-        catch(ClassCastException e)
-        { }
-    }
-
-    /**
-     * Only used in Aperture undo/redo system
-     * @return the value returned by this {@link #get()}
-     */
-    @Override
-    public Object getValue()
-    {
-        return this.get();
-    }
-
-    /**
-     * @return the default value that this type produces when not being initialised.
-     *         This is used in {@link #set(T)}, for example, to avoid null values for primitive datatype wrappers.
-     */
-    protected T getNullValue()
-    {
-        return null;
-    }
-
-    /**
-     * @return a deep copy of this object
-     */
-    @Override
-    public abstract GenericValue<T> copy();
-
-    /**
-     * Copy the {@link #value} from the specified object to this object.
-     * @param origin the origin that should be copied from
-     */
-    @Override
-    public abstract void copy(Value origin);
-
-    /**
-     * Compare the objects based on their {@link #value} variables. Ignores the other variables.
-     * @param obj
-     * @return true if this object's {@link #value} equals the specified object's {@link #value}, using {@link #equals(Object)}.
-     *         Or if the specified object or its value and this.value are both null.
-     */
-    @Override
-    public boolean equals(Object obj)
-    {
-        if (!(obj instanceof GenericValue))
-        {
-            /*
-             * if this.value == null and obj is null return true because both values are null
-             */
-            return (obj == null && this.value == null);
-        }
-
-        GenericValue valueObj = (GenericValue) obj;
-
-        if (valueObj.value == null && this.value == null)
-        {
-            return true;
-        }
-        else if (valueObj.value == null)
-        {
-            return false;
-        }
-        else if (valueObj.value.equals(this.value))
-        {
-            return true;
-        }
-
-        return this == obj;
     }
 
     /**
@@ -228,50 +81,4 @@ public abstract class GenericValue<T> extends Value
 
         return !this.value.equals(this.defaultValue);
     }
-
-
-    /**
-     * Read all contents from this object from the ByteBuffer and call {@link #superFromBytes(ByteBuf)}
-     * @param buffer
-     */
-    @Override
-    public abstract void fromBytes(ByteBuf buffer);
-
-    /**
-     * Write all contents from this object to the ByteBuffer and call {@link #superToBytes(ByteBuf)}
-     * @param buffer
-     */
-    @Override
-    public abstract void toBytes(ByteBuf buffer);
-
-    /**
-     * Only read the {@link #value} from the ByteBuffer
-     * @param buffer
-     */
-    public abstract void valueFromBytes(ByteBuf buffer);
-
-    /**
-     * Only write the {@link #value} into the ByteBuffer
-     * @param buffer
-     */
-    public abstract void valueToBytes(ByteBuf buffer);
-
-    @Override
-    public abstract void valueFromJSON(JsonElement element);
-
-    @Override
-    public abstract JsonElement valueToJSON();
-
-    /**
-     * set the value based on the specified tag
-     * @param tag the tag should be the value without the necessity to search for a key
-     */
-    public abstract void valueFromNBT(NBTBase tag);
-
-    /**
-     * @return the value as instance of a subclass of {@link net.minecraft.nbt.NBTBase}.
-     *         It can also return null, depending on the implementation in the subclasses.
-     */
-    @Nullable
-    public abstract NBTBase valueToNBT();
 }
