+++ src/main/java/mchorse/mclib/utils/ValueSerializer.java
@@ -3,6 +3,7 @@ package mchorse.mclib.utils;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import io.netty.buffer.ByteBuf;
+import mchorse.mclib.config.values.GenericBaseValue;
 import mchorse.mclib.config.values.GenericValue;
 import mchorse.mclib.network.IByteBufSerializable;
 import mchorse.mclib.network.INBTSerializable;
@@ -10,8 +11,10 @@ import net.minecraft.nbt.NBTBase;
 import net.minecraft.nbt.NBTTagCompound;
 
 import javax.annotation.Nullable;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
@@ -23,7 +26,9 @@ import java.util.UUID;
  * Ideally, the references to these instances will not change throughout their lifetime, so the usage of final is advised.
  * In the constructor of the class with the variables, the references to all the GenericValue variables 
  * can now be registered to a ValueSerializer instance, 
- * using {@link #registerValue(String, String, GenericValue)} or {@link #registerJSONValue(String, GenericValue)} or {@link #registerNBTValue(String, GenericValue)}.
+ * using {@link #registerValue(String, String, GenericBaseValue)} or {@link #registerJSONValue(String, GenericBaseValue)} or {@link #registerNBTValue(String, GenericBaseValue)}.
+ *
+ * @author Christian F. (Chryfi)
  */
 public class ValueSerializer implements IByteBufSerializable, INBTSerializable
 {
@@ -31,27 +36,51 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
      * The key is a UUID String.
      * This HashMap will not include duplicate GenericValue references.
      */
-    private final Map<String, GenericValue<?>> pool = new LinkedHashMap<>();
+    private final Map<String, GenericBaseValue<?>> pool = new LinkedHashMap<>();
 
     /**
      * Key is the name that is used for serialization/deserialization
      * Value is the UUID String pointing to the key of {@link #pool}
      */
     private final Map<String, String> nbtMap = new HashMap<>();
+    /**
+     * Contains the UUID to a value if it should always be written, no matter if it changed or not
+     */
+    private final List<String> nbtAlwaysWrite = new ArrayList<>();
+
     private final Map<String, String> jsonMap = new HashMap<>();
+    /**
+     * Contains the UUID to a value if it should always be written, no matter if it changed or not
+     */
+    private final List<String> jsonAlwaysWrite = new ArrayList<>();
 
     /**
      * Register the provided GenericValue object reference
      * with the provided key name for NBT serialization/deserialization.
      * If the provided value is null or if the key already exists, it will not be registered.
      *
-     * @param nbt name that should be used for serialization/deserialization
+     * @param nbt name that should be used for serialization/deserialization.
      *            If it is empty, the value will not be registered for NBT serialization/deserialization
      * @param value the reference to the value that should be serialization/deserialization
      */
-    public void registerNBTValue(String nbt, GenericValue<?> value)
+    public void registerNBTValue(String nbt, GenericBaseValue<?> value)
     {
-        this.registerValue(nbt, "", value);
+        this.registerValue(nbt, "", value, false, false);
+    }
+
+    /**
+     * Register the provided GenericValue object reference
+     * with the provided key name for NBT serialization/deserialization.
+     * If the provided value is null or if the key already exists, it will not be registered.
+     *
+     * @param nbt name that should be used for serialization/deserialization.
+     *            If it is empty, the value will not be registered for NBT serialization/deserialization
+     * @param value the reference to the value that should be serialization/deserialization
+     * @param alwaysWrite when true, the value will always be serialized to NBT, no matter if it changed or not.
+     */
+    public void registerNBTValue(String nbt, GenericBaseValue<?> value, boolean alwaysWrite)
+    {
+        this.registerValue(nbt, "", value, alwaysWrite, false);
     }
 
     /**
@@ -59,13 +88,28 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
      * with the provided key name for JSON serialization/deserialization.
      * If the provided value is null or if the key already exists, it will not be registered.
      *
-     * @param json name that should be used for serialization/deserialization
+     * @param json name that should be used for serialization/deserialization.
      *                If it is empty, the value will not be registered for JSON serialization/deserialization
      * @param value the reference to the value that should be serialization/deserialization
      */
-    public void registerJSONValue(String json, GenericValue<?> value)
+    public void registerJSONValue(String json, GenericBaseValue<?> value)
     {
-        this.registerValue("", json, value);
+        this.registerValue("", json, value, false, false);
+    }
+
+    /**
+     * Register the provided GenericValue object reference
+     * with the provided key name for JSON serialization/deserialization.
+     * If the provided value is null or if the key already exists, it will not be registered.
+     *
+     * @param json name that should be used for serialization/deserialization.
+     *                If it is empty, the value will not be registered for JSON serialization/deserialization
+     * @param value the reference to the value that should be serialization/deserialization
+     * @param alwaysWrite when true, the value will always be serialized to JSON, no matter if it changed or not.
+     */
+    public void registerJSONValue(String json, GenericBaseValue<?> value, boolean alwaysWrite)
+    {
+        this.registerValue("", json, value, false, alwaysWrite);
     }
 
     /**
@@ -79,18 +123,44 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
      *                If it is empty, the value will not be registered for JSON serialization/deserialization
      * @param value the reference to the value that should be serialization/deserialization
      */
-    public void registerValue(String nbt, String json, GenericValue<?> value)
+    public void registerValue(String nbt, String json, GenericBaseValue<?> value)
+    {
+        this.registerValue(nbt, json, value, false, false);
+    }
+
+    /**
+     * Register the provided GenericValue object reference
+     * with the provided key names for NBT and JSON serialization/deserialization.
+     * If the provided value is null or if the key already exists, it will not be registered.
+     *
+     * @param nbt name that should be used for serialization and deserialization.
+     *            If it is empty, the value will not be registered for NBT serialization/deserialization
+     * @param json name that should be used for serialization/deserialization
+     *                If it is empty, the value will not be registered for JSON serialization/deserialization
+     * @param value the reference to the value that should be serialization/deserialization
+     * @param alwaysWriteJSON when true, the value will always be serialized to JSON, no matter if it changed or not.
+     * @param alwaysWriteNBT when true, the value will always be serialized to NBT, no matter if it changed or not.
+     */
+    public void registerValue(String nbt, String json, GenericBaseValue<?> value, boolean alwaysWriteNBT, boolean alwaysWriteJSON)
     {
         if (value == null) return;
 
         if (!nbt.isEmpty() && !this.nbtMap.containsKey(nbt))
         {
-            this.nbtMap.put(nbt, this.poolValue(value));
+            String uuid = this.poolValue(value);
+
+            this.nbtMap.put(nbt, uuid);
+
+            if (alwaysWriteNBT) this.nbtAlwaysWrite.add(uuid);
         }
 
         if (!json.isEmpty() && !this.jsonMap.containsKey(json))
         {
-            this.jsonMap.put(json, this.poolValue(value));
+            String uuid = this.poolValue(value);
+
+            this.jsonMap.put(json, uuid);
+
+            if (alwaysWriteJSON) this.jsonAlwaysWrite.add(uuid);
         }
     }
 
@@ -100,7 +170,7 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
      * @return the new uuid string of the pooled value, or if the value already existed return the uuid of it.
      *         If the provided value was null, null will be returned.
      */
-    protected String poolValue(GenericValue<?> value)
+    protected String poolValue(GenericBaseValue<?> value)
     {
         if (value == null) return null;
 
@@ -123,11 +193,11 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
      *         Returns null if the reference has not been found.
      */
     @Nullable
-    protected String getValueReference(GenericValue<?> value)
+    protected String getValueReference(GenericBaseValue<?> value)
     {
         if (value == null) return null;
 
-        for (Map.Entry<String, GenericValue<?>> entry : this.pool.entrySet())
+        for (Map.Entry<String, GenericBaseValue<?>> entry : this.pool.entrySet())
         {
             if (entry.getValue() == value)
             {
@@ -139,38 +209,38 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
     }
 
     /**
-     * Calls {@link GenericValue#valueFromBytes(ByteBuf)}
+     * Calls {@link GenericBaseValue#valueFromBytes(ByteBuf)}
      */
     @Override
     public void fromBytes(ByteBuf buffer)
     {
-        for (GenericValue<?> value : this.pool.values())
+        for (GenericBaseValue<?> value : this.pool.values())
         {
             value.valueFromBytes(buffer);
         }
     }
 
     /**
-     * Calls {@link GenericValue#valueToBytes(ByteBuf)}
+     * Calls {@link GenericBaseValue#valueToBytes(ByteBuf)}
      */
     @Override
     public void toBytes(ByteBuf buffer)
     {
-        for (GenericValue<?> value : this.pool.values())
+        for (GenericBaseValue<?> value : this.pool.values())
         {
             value.valueToBytes(buffer);
         }
     }
 
     /**
-     * Calls {@link GenericValue#valueFromNBT(NBTBase)}
+     * Calls {@link GenericBaseValue#valueFromNBT(NBTBase)}
      */
     @Override
     public void fromNBT(NBTTagCompound tag)
     {
         for (Map.Entry<String, String> entry : this.nbtMap.entrySet())
         {
-            GenericValue value = this.pool.get(entry.getValue());
+            GenericBaseValue value = this.pool.get(entry.getValue());
             String key = entry.getKey();
 
             if (tag.hasKey(key))
@@ -181,27 +251,31 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
     }
 
     /**
-     * Calls {@link GenericValue#valueToNBT()}
+     * Calls {@link GenericBaseValue#valueToNBT()}. Only serializes the value if it has changed.
+     * Serializes always, if the value has been registered with alwaysWrite flag true.
      */
     @Override
     public NBTTagCompound toNBT(NBTTagCompound tag)
     {
         for (Map.Entry<String, String> entry : this.nbtMap.entrySet())
         {
-            GenericValue value = this.pool.get(entry.getValue());
-            String key = entry.getKey();
+            GenericBaseValue value = this.pool.get(entry.getValue());
+            String uuid = entry.getKey();
 
-            if (value.hasChanged() && value.get() != null)
+            if (!this.nbtAlwaysWrite.contains(uuid) && value instanceof GenericValue && !((GenericValue) value).hasChanged())
             {
-                tag.setTag(key, value.valueToNBT());
+                continue;
             }
+
+            tag.setTag(uuid, value.valueToNBT());
         }
 
         return tag;
     }
 
     /**
-     * Calls {@link GenericValue#valueToJSON()}
+     * Calls {@link GenericBaseValue#valueToJSON()}. Only serialize the value if it has changed.
+     * Serializes always, if the value has been registered with alwaysWrite flag true.
      */
     public JsonElement toJSON()
     {
@@ -209,20 +283,22 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
 
         for (Map.Entry<String, String> entry : this.jsonMap.entrySet())
         {
-            GenericValue value = this.pool.get(entry.getValue());
-            String key = entry.getKey();
+            GenericBaseValue value = this.pool.get(entry.getValue());
+            String uuid = entry.getKey();
 
-            if (value.hasChanged() && value.get() != null)
+            if (!this.jsonAlwaysWrite.contains(uuid) && value instanceof GenericValue && !((GenericValue) value).hasChanged())
             {
-                jsonRoot.add(key, value.valueToJSON());
+                continue;
             }
+
+            jsonRoot.add(uuid, value.valueToJSON());
         }
 
         return jsonRoot;
     }
 
     /**
-     * Calls {@link GenericValue#valueFromJSON(JsonElement)}
+     * Calls {@link GenericBaseValue#valueFromJSON(JsonElement)}
      * @param element
      */
     public void fromJSON(JsonElement element)
@@ -236,7 +312,7 @@ public class ValueSerializer implements IByteBufSerializable, INBTSerializable
 
         for (Map.Entry<String, String> entry : this.jsonMap.entrySet())
         {
-            GenericValue value = this.pool.get(entry.getValue());
+            GenericBaseValue value = this.pool.get(entry.getValue());
             String key = entry.getKey();
 
             if (jsonObject.has(key))
